remove fem stub, add friendly test, revert error message test
